+++ cmd/migrate.go
@@ -66,18 +66,20 @@ var migrateCmd = &cobra.Command{
 		s := store.NewWithClient(r)
 
 		// Login to the remote database
-		token, err := s.Login(username, password)
+		slog.Debug("logging in", "username", username, "password", "[REDACTED]")
+		loginResponse, err := r.R().SetBody(map[string]interface{}{"username": username, "password": password}).SetResult(&store.Token{}).Post("/auth/login")
 		if err != nil {
 			slog.Error("could not login", "error", err)
 			return err
 		}
-		if token == "" {
+		token := loginResponse.Result().(*store.Token)
+		if token.AccessToken == "" {
 			slog.Error("no token returned")
 			return err
 		}
 
-		// Create a new authenticated HTTP client and set it on the store
-		r.SetAuthToken(token)
+		// Set the auth token
+		r.SetAuthToken(token.AccessToken)
 
 		// Set the work item status to 'migrating'
 		response, err := s.UpdateWorkItem(*item, store.MIGRATING)
