--- internal/store/auth.go
@@ -1,10 +1,0 @@
-package store
-
-type Credentials struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
-
-type Token struct {
-	AccessToken string `json:"access_token"`
-}
