+++ internal/store/token.go
@@ -1,0 +1,5 @@
+package store
+
+type Token struct {
+	AccessToken string `json:"access_token"`
+}
